+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.0.0
+
+- added dependency on `DerailValleyModToolbar`
+- removed unused dependency `CommsRadioAPI`
+- added "derail train" button
+- added "prevent derail" checkbox
+
 # 0.0.1
 
 Initial version.